rag-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
qdrant-client==1.7.3

# AI/ML
openai==1.10.0

# Utilities
python-multipart==0.0.6
